Added CreateImageReader methods to IProcessorArchitecture to make creation of reads platform-agnostic.
Implemented opcode 0x82 in X86Disassembler.

Extended the Configuration/LoaderElement class to support file name extensions as well as "magic numbers".

Fixed bug in Constant.ToByte()

Several modifications to the Project model.

Fixed bugs in LookaheadEnumerator class.

Start of support for adding metadata files to a project, that is files that are used solely for type information.

Continued work on making the DisassemblyControl using TextModels.

Fixed a bug in the regexp engine for searching after binary patterns.