IProcessorArchitecture.GetAddressOffset returns the offset part of an address. Thank you intel for x86 segmented addresses.
Refactored X86PointerScanner.

Renamed ProcedureSignature.ArgumentsValid to ParametersValid.

Fixed a bug where the Backwalker was returning Register.None as the index register and then passing that on to the environment.

Made ProgramGraph a public class.

AddressSearchResults should be sortable.

Removed all the properties of Scanner and IScanners that already exist on Program