Added SysV project - the default ABI for ELF files.
Added moved and punpck* instuctions to the x86 disassembler and rewriter.

Fixed a couple of bugs in the C => XML converter.

Bumped version numbers to 0.4.2.0 in preparation for new release.

Added Segment renderers for ELF .rel and .symtab segments.

More work to get a proper TypeLib loader up and running.