Added an `Architecture` property to Procedure
Previously, all procedures were assumed to be of the same architecture as the Program. However, it turns out that many executables, such as ARM ELF binaries, support scenarios where more than one architecture is supported. Specifically, there are binaries that contain both 32-bit ARM and 16-bit ARM Thumb instructions.

In order to support this, the notion of "current processor architecture" has to be moved from Program into each Procedure. Later commits will make the Scanner sensitive to a "current processur architecture" when architecture switches occur. This could happen with the ARM BX instruction, or in the case of BASIC programs, calling SYS 2048, which calls the machine code at address 2048.