Fix: DisassemblyTextModel
The DisassemblyTextModel was having problems displaying 16-bit address spaces. The issue is that the EndPosition of the text model needs to be one past the last representable address (0xFFFF). But this position would wrap around to (0x0000) because it overflows the 16-bits, and we end up with EndPosition == 0x0000. The fix replaces Addresses in the text model with long offsets.