Refactor: remove mutable Expression.TypeVariable.
This mutable property is only used during type inference. Instead, an Expression -> TypeVariable mapping is maintained in
TypeStore.

The changes in the output are due to the fact that the `Address` expression subclass has its own implementation of GetHashCode(). When two separate instances of the Address class are placed in the TypeStore.mapExprToTypevar dictionary, they will end up sharing the same TypeVariable. This is acceptable: a global pointer is a global pointer and it should have a single type variable.