AEON: Add bt.lwst instruction
The bt.lwst instruction uses almost the same encoding as bt.swst, except
it has the LSB set. This means the immediate actually starts at bit 1,
is 4 bits long, and is shifted left by 2 bits to calculate the offset.

Note that both the bt.lwst and bt.swst mnemonics are guesses not based
on any official documentation.