Fix: stop throwing exception in TypedExpressionRewriter
One of the most common exceptions Reko was raising was caused by
expressions like `a + b`, where both ``[[a]]` and `[[b]]` were "complex"
types -- i.e. not primitives. These are almost always caused by failures
earlier in the decompilation pipeline.

Rather than throwing exceptions, we emit the code, with the hope that
future commits fix the problems causing the complex types to appear.