Fix: exception records in PE Images
PE RUNTIME_FUNCTION structures in the exception section of X86-64 executables
don't always point to the beginning of procedures. If the UNWIND_INFO Flags
indicate that the RUNTIME_FUNCTION structure is UNW_FLAG_CHAININFO, then
the "function start address" may actually be pointing to any code block inside
the procedure, and cannot be used as a pointer to a procedure.