Feature: adjust out registers based on liveness from callers.
Reko detects what registers are live out of procedures by examining all
callers to determine if said callers are using the registers. Now, we also
reduce the size of the register being used. This handles the case where
the CPU architecture has 32-bit registers, but the value being returned
is smaller.